tests/requirements-fl-test.txt
# Test requirements for federated learning module
# Install with: uv pip install -r tests/requirements-fl-test.txt

# Core testing framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.1

# ML/DL frameworks (CPU versions for testing)
torch>=2.0.0
torchvision>=0.15.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Flower federated learning
flwr>=1.5.0

# Additional utilities
pillow>=10.0.0
pyyaml>=6.0
